Moved collapse_nphc in preclassical
